requirements.txt
fastapi
uvicorn[standard]
pandas
pydantic
playwright
google-genai
Pillow 